requirements.txt
-i https://pypi.org/simple
asgiref==3.3.4
dj-database-url==0.5.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django==3.2.4
djangorestframework-simplejwt==4.7.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.4.0 ; python_version < '3.8'
markdown==3.3.4
pillow==8.2.0
psycopg2==2.8.6
pyjwt==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0 ; python_version < '3.8'
whitenoise==5.2.0
zipp==3.4.1
